Add a separate core option for people who don't want jsonify playfill
This is a non-breaking change that allows people to optionally
`require(‘json-stable-stringify/core’)` and save roughly 4.5KB (before
Gzip)
